requirements.txt
flask
mysqlclient
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
flask-cors
dotenv

scikit-learn
matplotlib
joblib
numpy